refactor blocked component into own file
